misc fixes as requested in review, newlines added, changes routes, added redirects
