+++ api/src/router/admin.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { NotAuthorizedError } from "../errors";
-import User, { UserType } from "../models/user";
+import { UserType } from "../models/user";
 
 import config, { configInterface } from "../config/config";
 import { protectedRoute } from "../middleware/protected-route";
@@ -43,51 +43,11 @@ const updateSettings = async (
   }
 };
 
-const getUsersPermissions = async (
-  _: express.Request,
-  res: express.Response,
-  next: express.NextFunction
-) => { 
-  const user = res.locals.user;
-
-  try {
-    if (user.role !== UserType.COORDINATOR) throw new NotAuthorizedError();
-    const users = await User.find();
-    res.json({
-      status: 200, 
-      message: "Successfully fetched users",
-      data: users
-    });
-  } catch (err) {
-    next(err);
-  }
-};
-
-const changeUserPermission = async (
-  req: express.Request,
-  res: express.Response,
-  next: express.NextFunction
-) => {
-  const user = res.locals.user;
-  try {
-    if (user.role !== UserType.COORDINATOR) throw new NotAuthorizedError();
-    User.update(req.body.id, req.body.partial)
-    res.json({
-      status: 200,
-      message: "Succesfully changed user permission",
-    })
-  } catch (err) {
-    next(err);
-  }
-}
-
 const adminRouter = express.Router();
 adminRouter.use(protectedRoute);
 
 adminRouter.get("/settings", reqUser, getSettings);
 adminRouter.post("/settings", reqUser, updateSettings);
-adminRouter.get("/users", reqUser, getUsersPermissions)
-adminRouter.patch("/users", reqUser, changeUserPermission)
 
 
 export default adminRouter;
