+++ ui/styles/permissions.module.css
